refactor: move get submission to staff area
